frontend/requirements.txt
streamlit
pillow
fastapi
